Refactor the 'getModel' callbacks into their own file
