+++ cmd/frontend/internal/httpapi/completions/chat.go
@@ -5,17 +5,9 @@ import (
 
 	"net/http"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/cody"
-	sgactor "github.com/sourcegraph/sourcegraph/internal/actor"
-	"github.com/sourcegraph/sourcegraph/internal/dotcom"
-
 	"github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/internal/completions/client/anthropic"
-	"github.com/sourcegraph/sourcegraph/internal/completions/client/fireworks"
-	"github.com/sourcegraph/sourcegraph/internal/completions/client/google"
 	"github.com/sourcegraph/sourcegraph/internal/completions/types"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/redispool"
 	"github.com/sourcegraph/sourcegraph/internal/telemetry/telemetryrecorder"
@@ -43,79 +35,5 @@ func NewChatCompletionsStreamHandler(logger log.Logger, db database.DB) http.Han
 		types.CompletionsFeatureChat,
 		rl,
 		"chat",
-		func(ctx context.Context, requestParams types.CodyCompletionRequestParameters, c *conftypes.CompletionsConfig) (string, error) {
-			// Allow a number of additional models on Dotcom
-			if dotcom.SourcegraphDotComMode() {
-				actor := sgactor.FromContext(ctx)
-				user, err := actor.User(ctx, db.Users())
-				if err != nil {
-					return "", err
-				}
-
-				subscription, err := cody.SubscriptionForUser(ctx, db, *user)
-				if err != nil {
-					return "", err
-				}
-
-				if isAllowedCustomChatModel(requestParams.Model, subscription.ApplyProRateLimits) {
-					return requestParams.Model, nil
-				}
-			}
-			// No user defined models for now.
-			if requestParams.Fast {
-				return c.FastChatModel, nil
-			}
-			return c.ChatModel, nil
-		},
-	)
-}
-
-// We only allow dotcom clients to select a custom chat model and maintain an allowlist for which
-// custom values we support
-func isAllowedCustomChatModel(model string, isProUser bool) bool {
-	// When updating these two lists, make sure you also update `allowedModels` in codygateway_dotcom_user.go.
-	if isProUser {
-		switch model {
-		case
-			"anthropic/" + anthropic.Claude3Haiku,
-			"anthropic/" + anthropic.Claude3Sonnet,
-			"anthropic/" + anthropic.Claude3Opus,
-			"fireworks/" + fireworks.Mixtral8x7bInstruct,
-			"fireworks/" + fireworks.Mixtral8x22Instruct,
-			"openai/gpt-3.5-turbo",
-			"openai/gpt-4o",
-			"openai/gpt-4-turbo",
-			"openai/gpt-4-turbo-preview",
-			"google/" + google.Gemini15FlashLatest,
-			"google/" + google.Gemini15ProLatest,
-			"google/" + google.GeminiProLatest,
-			"google/" + google.Gemini15Flash,
-			"google/" + google.Gemini15Pro,
-			"google/" + google.GeminiPro,
-
-			// Remove after the Claude 3 rollout is complete
-			"anthropic/claude-2",
-			"anthropic/claude-2.0",
-			"anthropic/claude-2.1",
-			"anthropic/claude-instant-1.2-cyan",
-			"anthropic/claude-instant-1.2",
-			"anthropic/claude-instant-v1",
-			"anthropic/claude-instant-1":
-			return true
-		}
-	} else {
-		switch model {
-		case
-			"anthropic/" + anthropic.Claude3Haiku,
-			"anthropic/" + anthropic.Claude3Sonnet,
-			// Remove after the Claude 3 rollout is complete
-			"anthropic/claude-2",
-			"anthropic/claude-2.0",
-			"anthropic/claude-instant-v1",
-			"anthropic/claude-instant-1":
-			return true
-		}
-	}
-
-	return false
+		getChatModelFn(db))
 }
